+++ tools/__init__.py
@@ -1,6 +1,7 @@
 from .send_message		import send_message_to_QQ
 from .download_file		import download_qq_file
 from .query_message_id	import query_qq_message_id
+from .delete_message	import delete_qq_message
 
 
 __all__ = [
@@ -8,5 +9,6 @@ __all__ = [
 	"send_message_to_QQ",
 	"download_qq_file",
 	"query_qq_message_id",
+	"delete_qq_message",
 
 ]